added scale thing and extracted class
